Also consider `rdf_[un]linked_shared_data` in `update_wikidata`
So far, the `update_wikidata` command only considered the parts
`rdf_added_data` (triples that must be added) and `rdf_deleted_data`
(triples that must be removed) of each message from the Wikidata update
stream.

Now also consider `rdf_linked_shared_data` (triples which might have
already been added and if not must be added) and
`rdf_unlinked_shared_data` (triples which should be removed, but do not
harm if they are not removed). See
https://wikitech.wikimedia.org/wiki/Wikidata_Query_Service/Streaming_Updater/Public_Update_Stream

NOTE: Messages where the value for `"operation"` is one of `"import"`,
`"diff"` or `"reconcile"` are now processed properly. However, those
where the value is `"delete"` are not (in that case, all four of the
`rdf_...` above are empty and a `DELETE WHERE` query would be needed.