Consider `delete` operations in Wikidata update stream
So far, the `delete` operations were simply ignored. Now, a batch is
finished as soon as a `delete` operation is encountered. However, this
impacts performance when there are many `delete` operations (larger
batches are better for performance).

TODO: Push delete operation to the end of a batch as much as possible
(that is, as long as no insertion triples involving the deleted entities
are encountered).