Now handle deletes both correctly and efficiently
The loop has now been refactored such that a new stream connection is
created for each batch. That way, whenever we read a message that we
don't want to process anymore in the current batch, we can just finished
the batch and resumt from that message in the next batch.

Delete operations are not postponed until the end of a batch as before.
And a batch is only finished prematurely if we encounter an operation
that wants to add data for an entity that has previously been deleted in
the same batch. This happens quite rarely in practice, so that most
batches reach their maximum size (until we approach the present time).