Use `offset` for everything now and introduce `--use-cached-sparql-queries`
So far, the `timestamp` of the event ID was used for resuming. But that
is inherently unreliable because there can be multiple events with the
same timestamp, so that when setting `Last-Event-ID` to the last seen
timestamp, we might miss some events. Instead, we now use the `offset`
throughout the codebase for resuming.

The SPARQL queries are now written to files with the name
`update.<offset>.<batch-size>.sparql` (where `<batch-size>` is the
number of messages actually processed in that batch, which can be
smaller then the target batch size) instead of the previous
`update.sparql.<index>`. This allows reusing those files with the new
option `--use-cached-sparql-queries`, which saves re-processing of the
respecitve messages (which takes longer than the update itself).