First version of `rebuid-index` command
Rebuild a fresh index from the current data (including update triples).
This is much faster than first exporting the complete data and
rebuilding the index from that.

For example, for Wikidata (with around 18 billion triples),
`rebuid-index` takes only around half an hour. For DBLP (with around
500 million triples), it takes less than one minute.

NOTE: This currently requires the adfreiburg/qlever:REBUILD-INDEX Docker
image, which is still in BETA.